+++ guix_env/template_scripts/run_script.sh
@@ -3,8 +3,9 @@
 # Link the lib file for FHS library handling
 export LD_LIBRARY_PATH=/lib
 
-rm  ${HOME}/.guix_env/guix_env_venv/{{ name }}_venv/bin/python3
-ln -s $(which python3) ${HOME}/.guix_env/guix_env_venv/{{ name }}_venv/bin/python3
-source ${HOME}/.guix_env/guix_env_venv/{{ name }}_venv/bin/activate
+[ -f ${HOME}/.guix_env/guix_env_venv/{{ name }}_venv/bin/python3 ] && rm  ${HOME}/.guix_env/guix_env_venv/{{ name }}_venv/bin/python3 && ln -s $(which python3) ${HOME}/.guix_env/guix_env_venv/{{ name }}_venv/bin/python3
 
-$@
+export SHELL=$(realpath $(which zsh))
+export ZDOTDIR=${HOME}/.guix_env/{{ name }}/bin
+
+cd ${HOME}/.guix_env/{{ name }} && poetry run $@
